fix(observability): single-source build id so the header can't diverge from /health (Codex F12)
Three resolvers computed the build identity independently — health.ts and
v2/run.ts held near-identical copies, and the onSend x-olumi-service-build
header used its own inline BUILD_ID||GITHUB_SHA||'dev' with no git fallback.
On a deploy where neither env var is set the header emitted 'dev' while
/health reported the real short SHA. Extract one canonical getBuildId() into
src/util/build-id.ts; health, run, and the header hook all use it. Pin the
header==body invariant with a RED-first test.
